requirements.txt
eth-brownie==1.20.7
fastapi==0.115.8
uvicorn[standard]==0.34.0
web3==7.8.0
pymongo==4.11.1
python-dotenv==1.0.1
pydantic==2.10.6
requests==2.32.3
pytest==8.3.4
